Create CCSSNsearch.ps1
this file will search a windows machine for files that contain credit card numbers or social security numbers.  if it finds a number it will create a custom Windows Event ID of 10001 in the Application log.  use this for data loss prevention in conjunction with scheduled tasks and filter for the event ID at your SIEM.